src/ACMESky/camunda/requirements.txt
annotated-types==0.7.0
appdirs==1.4.4
attrs==23.2.0
cached-property==1.5.2
camunda-external-task-client-python3==4.4.0
certifi==2024.2.2
chardet==5.2.0
charset-normalizer==3.3.2
defusedxml==0.7.1
frozendict==2.4.4
greenlet==3.0.3
idna==3.7
isodate==0.6.1
javaobj-py3==0.4.4
lxml==5.2.2
platformdirs==4.2.2
psycopg2==2.9.9
pydantic==2.7.1
pydantic_core==2.18.2
pytz==2024.1
requests==2.32.2
requests-file==2.1.0
requests-toolbelt==1.0.0
six==1.16.0
SQLAlchemy==2.0.30
typing_extensions==4.12.0
urllib3==2.2.1
zeep==4.2.1
